requirements.txt
pillow
black
opencv-python
#schließt numpy mit ein
numpy